backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
aiosqlite>=0.19.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
bcrypt>=5.0.0
python-multipart>=0.0.6
redis>=5.0.1
python-dotenv>=1.0.0
pytest
httpx
strands-agents[litellm]>=0.1.0
strands-agents-tools>=0.1.0
